refactor: 도메인이 없는 공통 훅을 hooks/common 으로 이동
